gameserver/example/requirements.txt
websocket-client
google
protobuf
